feat: 🚀 actualizar el tipo de dato del campo 'amount' en los DTOs y entidades de venta a 'float8' y 'IsNumber' para permitir valores decimales y mejorar la validación

+++ src/sales/entities/sale-details.entity.ts
@@ -31,7 +31,7 @@ export class SaleDetails {
   unit_of_measure: MassUnit;
 
   @Column({
-    type: 'int4',
+    type: 'float8',
   })
   amount: number;
